backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-dotenv==1.0.1
boto3==1.35.33
Pillow==11.1.0
python-multipart==0.0.12
pydantic==2.10.6
